+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix for "upgrade all packages"
 - Fix upgrade to version of package if package is on latest
 - Fix when esc is pressed when selecting device
+- Better error message when shell is not set to /bin/zsh
 
 ### Version 1.3.1
 - Mono Repo Support for Nx